MA-17: refactoring FLUX system, rendering
