+++ src/Redux/State.js
@@ -1,4 +1,6 @@
-import { rerenderEntireTree } from '../render'
+let rerenderEntireTree = () => {
+	console.log('state was changed')
+}
 
 let state = {
 	profilePage: {
@@ -25,7 +27,7 @@ let state = {
 	},
 }
 
-export let addPost = () => {
+export const addPost = () => {
 	let newPost = {
 		id: 5,
 		message: state.profilePage.newPostText,
@@ -36,9 +38,12 @@ export let addPost = () => {
 	rerenderEntireTree(state)
 }
 
-export let updateNewPostText = newText => {
+export const updateNewPostText = newText => {
 	state.profilePage.newPostText = newText
 	rerenderEntireTree(state)
 }
 
+export const subscribe = observer => {
+	rerenderEntireTree = observer
+}
 export default state
