+++ src/index.js
@@ -1,5 +1,29 @@
 import './index.css'
-import state from './Redux/State'
-import { rerenderEntireTree } from './render'
+import state, { subscribe } from './Redux/State'
+import React from 'react'
+import ReactDOM from 'react-dom/client'
+import './index.css'
+import reportWebVitals from './reportWebVitals'
+import App from './App'
+import { addPost, updateNewPostText } from './Redux/State'
+import { BrowserRouter } from 'react-router-dom'
+
+const root = ReactDOM.createRoot(document.getElementById('root'))
+
+export let rerenderEntireTree = state => {
+	root.render(
+		<React.StrictMode>
+			<BrowserRouter>
+				<App
+					state={state}
+					addPost={addPost}
+					updateNewPostText={updateNewPostText}
+				/>
+			</BrowserRouter>
+		</React.StrictMode>
+	)
+}
+reportWebVitals()
 
 rerenderEntireTree(state)
+subscribe(rerenderEntireTree)
