--- src/render.js
@@ -1,24 +1,0 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import './index.css'
-import reportWebVitals from './reportWebVitals'
-import App from './App'
-import { addPost, updateNewPostText } from './Redux/State'
-import { BrowserRouter } from 'react-router-dom'
-
-const root = ReactDOM.createRoot(document.getElementById('root'))
-
-export let rerenderEntireTree = state => {
-	root.render(
-		<React.StrictMode>
-			<BrowserRouter>
-				<App
-					state={state}
-					addPost={addPost}
-					updateNewPostText={updateNewPostText}
-				/>
-			</BrowserRouter>
-		</React.StrictMode>
-	)
-}
-reportWebVitals()
